:lock: Added Private Routes
