+++ src/HOC/PrivateRoute.js
@@ -1,0 +1,7 @@
+import { Outlet, Navigate } from 'react-router-dom'
+
+const PrivateRoute = ({ auth }) => {
+    return auth ? <Outlet /> : <Navigate to='/login' />
+}
+
+export default PrivateRoute
